requirements.txt
sentence-transformers==3.4.1
transformers==4.48.3
joblib==1.3.2
matplotlib==3.10.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.3
pillow==11.1.0
openpyxl==3.1.5
scikit_learn==1.6.1
scipy==1.15.1
spacy==3.6.1
streamlit==1.31.0
torch==2.6.0
torchvision==0.21.0
openai-clip==1.0.1
